// frontend-admin/src/api/generated/model/receiptReprintRequest.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Kasse API
 * Registrierkasse API - RKSV uyumlu kasa sistemi
 * OpenAPI spec version: v1
 */

export interface ReceiptReprintRequest {
  /** @nullable */
  deviceId?: string | null;
  /** @nullable */
  idempotencyKey?: string | null;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  printerProfileId?: string | null;
  /** @nullable */
  reasonDetail?: string | null;
  /** @nullable */
  reprintReasonCode?: string | null;
}